local_ai_file_agent/requirements.txt
requests
pandas
openpyxl
python-docx
pypdf
reportlab
PyYAML
tabulate
lxml
